% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_alternate_layout.R
\name{add_alternate_layout.gridlayout}
\alias{add_alternate_layout.gridlayout}
\title{Add alternate layout options}
\usage{
\method{add_alternate_layout}{gridlayout}(
  layout,
  alternate_layout,
  lower_bound_width = NULL,
  upper_bound_width = NULL,
  container_height = NULL
)
}
\arguments{
\item{alternate_layout}{A \code{gridlayout} object or layout-as-markdown-table
defining layout. The elements in this layout must match those in your main
layout.}

\item{lower_bound_width}{Width in pixels above which this layout is used}

\item{upper_bound_width}{Width in pixels below which this layout is used}

\item{container_height}{How tall the layout should be. If left as \code{NULL} this
will simply inherit the main layout's height. Often this should be set to
\code{"auto"} to allow mobile user's to scroll. Beware though, you will want to
absolutely size any plot outputs in these cases.}
}
\description{
Adds a layout that will be triggered based upon the size of the viewport.
This is typically used to add a mobile view (or a desktop view if your app is
mobile first.)
}
