% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_plot.R
\name{card_plot_output}
\alias{card_plot_output}
\title{Card-container plot output}
\usage{
card_plot_output(
  outputId,
  click = NULL,
  dblclick = NULL,
  hover = NULL,
  brush = NULL,
  height = NULL,
  stretch = TRUE,
  ...
)
}
\arguments{
\item{stretch}{Set to \code{TRUE} if this \code{card_body} is eager to use any extra
vertical space is available in the card.}

\item{...}{Named arguments become attributes on the div containing the plot.}

\item{class}{Additional CSS classes to include on the card div.}
}
\description{
A card-aware wrapper of \code{shiny::plotOutput} that has smart defaults for
sizing
}
\seealso{
\link{grid_plot}
}
