% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex_stack.R
\name{flex_stack}
\alias{flex_stack}
\title{Div with vertically stacked elements}
\usage{
flex_stack(..., title, collapsible, scrollable, item_gap = "10px")
}
\arguments{
\item{...}{Arguments (typically children) passed to the \code{htmltools::div()}
that contain the card's contents}

\item{title}{Character string to go across top of panel with label. If left
blank the card contents will take up entire space.}

\item{collapsible}{Should the card be able to be collapsed (\code{TRUE} or
\code{FALSE})? Gridlayout will only show collapser if the layout allows it
(panel is entirely positioned within "auto" sized rows, and has a title).
Setting this to \code{FALSE} will mean collapsibility of the panel will never be
enabled, regardless of layout.}

\item{scrollable}{Should scroll-bars be added so content that is larger than
the panel can be seen?}

\item{item_gap}{How much space should there be between consecutive items?}
}
\value{
Elements from \code{...} wrapped in a \code{shiny::div()} with styles for
vertical stacking applied.
}
\description{
Contain a series of ui elements with each vertically stacked on top of
eachother. Internally uses css flexbox to align items.
}
\seealso{
\link{grid_card}
}
\keyword{internal}
