% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elements.R
\name{get_element_ids}
\alias{get_element_ids}
\title{Get ids of element in layout}
\usage{
get_element_ids(layout)
}
\arguments{
\item{layout}{Object of class \code{"gridlayout"}.}
}
\value{
Character vector of ids of all elements in layout
}
\description{
Get ids of element in layout
}
\examples{

grid_obj <- md_to_gridlayout(
  layout_table = "
        |      |120px   |1fr    |1fr    |
        |------|--------|-------|-------|
        |100px |header  |header |header |
        |1fr   |sidebar |plot_a |plot_a |
        |1fr   |sidebar |plot_b |plot_c |"
 )

get_element_ids(grid_obj)

}
\seealso{
\link{get_elements}
}
