% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elements.R
\name{get_elements}
\alias{get_elements}
\title{Get elements and positions out of a grid layout}
\usage{
get_elements(layout)
}
\arguments{
\item{layout}{Object of class `"gridlayout"`.}
}
\value{
List of all unique elements in the layout and their `id`,
  `{start,end}_row`, and `{start,end}_col`. Positions are indexed starting at
  1
}
\description{
Get elements and positions out of a grid layout
}
