% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_plot.R
\name{grid_card_plot}
\alias{grid_card_plot}
\title{Grid-positioned plot output}
\usage{
grid_card_plot(area, outputId = area, ..., has_border = TRUE)
}
\arguments{
\item{area}{Name of grid area, should match an area defined in the layout
section of the wrapping \code{grid_page()} or \code{grid_container()}.}

\item{outputId}{Output id of the plot output. Used to link to server code
generating plot. If left unset this will use the same value as the \code{area}
argument.}

\item{...}{
  Arguments passed on to \code{\link[=card_plot_output]{card_plot_output}}
  \describe{
    \item{\code{height}}{height in valid css units (see \code{\link[htmltools:validateCssUnit]{htmltools::validateCssUnit()}}
for more details.) Most use-cases will leave this unset and the plot will
fill the card as best it can.}
    \item{\code{stretch}}{Set to \code{TRUE} if this \code{card_body} is eager to use any extra
vertical space is available in the card.}
    \item{\code{click}}{This can be \code{NULL} (the default), a string, or an object
created by the \code{\link[shiny:clickOpts]{clickOpts()}} function. If you use a value like
\code{"plot_click"} (or equivalently, \code{clickOpts(id="plot_click")}),
the plot will send coordinates to the server whenever it is clicked, and
the value will be accessible via \code{input$plot_click}. The value will be
a named list  with \code{x} and \code{y} elements indicating the mouse
position.}
    \item{\code{dblclick}}{This is just like the \code{click} argument, but for
double-click events.}
    \item{\code{hover}}{Similar to the \code{click} argument, this can be \code{NULL}
(the default), a string, or an object created by the
\code{\link[shiny:hoverOpts]{hoverOpts()}} function. If you use a value like
\code{"plot_hover"} (or equivalently, \code{hoverOpts(id="plot_hover")}),
the plot will send coordinates to the server pauses on the plot, and the
value will be accessible via \code{input$plot_hover}. The value will be a
named list with \code{x} and \code{y} elements indicating the mouse
position. To control the hover time or hover delay type, you must use
\code{\link[shiny:hoverOpts]{hoverOpts()}}.}
    \item{\code{brush}}{Similar to the \code{click} argument, this can be \code{NULL}
(the default), a string, or an object created by the
\code{\link[shiny:brushOpts]{brushOpts()}} function. If you use a value like
\code{"plot_brush"} (or equivalently, \code{brushOpts(id="plot_brush")}),
the plot will allow the user to "brush" in the plotting area, and will send
information about the brushed area to the server, and the value will be
accessible via \code{input$plot_brush}. Brushing means that the user will
be able to draw a rectangle in the plotting area and drag it around. The
value will be a named list with \code{xmin}, \code{xmax}, \code{ymin}, and
\code{ymax} elements indicating the brush area. To control the brush
behavior, use \code{\link[shiny:brushOpts]{brushOpts()}}. Multiple
\code{imageOutput}/\code{plotOutput} calls may share the same \code{id}
value; brushing one image or plot will cause any other brushes with the
same \code{id} to disappear.}
  }}

\item{has_border}{Should the card be surrounded by a border? Set to \code{FALSE}
to turn off.}
}
\value{
A grid panel filled with plot output
}
\description{
A light wrapper for \code{shiny::plotOutput()} that uses gridlayout-friendly
sizing defaults. For more control look at combining \code{grid_card()} with
\code{card_plot_output()}
}
\examples{
if (FALSE) {
library(gridlayout)
library(shiny)
library(bslib)

shinyApp(
  ui = grid_page(
    layout = c(
      "header  header",
      "sidebar distPlot"
    ),
    row_sizes = c("50px", "1fr"),
    col_sizes = c("200px", "1fr"),
    grid_card_text("header", "This is my header"),
    grid_card(
      "sidebar",
      title = "Settings",
      sliderInput("bins","Number of bins:", 1, 50, 30, width = "100\%")
    ),
    grid_plot("distPlot")
  ),
  server = function(input, output) {
    output$distPlot <- renderPlot({
      x    <- faithful[, 2]
      bins <- seq(min(x), max(x), length.out = input$bins + 1)
      hist(x, breaks = bins, col = 'darkgray', border = 'white')
    })
  }
)
}
}
\seealso{
\link{card_plot_output}
}
