% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_container.R
\name{grid_container}
\alias{grid_container}
\title{Build gridlayout container for specified layout}
\usage{
grid_container(id = "grid-container", layout, elements, full_height = TRUE)
}
\arguments{
\item{id}{ID unique to this container (not the HTML will be prefixed with
\verb{grid-})}

\item{layout}{Either a markdown table representation (see
\code{\link{md_to_gridlayout}}) or a \code{gridlayout} object defining the
desired layout for your Shiny app.}

\item{elements}{named list of the UI definitions that will be used to fill
all cells. Names must match those provided in \code{layout}.}

\item{full_height}{Should the grid-containing element be made as tall as
possible? Set to \code{FALSE} if the grid is contained within another element
and is not the whole application's UI. In this case the grid-container will take
the full height of containing element.}
}
\value{
A taglist with grid elements wrapped inside a container div of class \code{id}.
}
\description{
Builds a gridlayout within a div of specified id. Not typically called
directly but can be used to create nested grids
}
\examples{

if (interactive()) {
library(gridlayout)
library(shiny)
requireNamespace("bslib", quietly = TRUE)
my_layout <- "
|      |        |       |
|------|--------|-------|
|2rem  |200px   |1fr    |
|150px |header  |header |
|1fr   |sidebar |plot   |"

# The classic Geyser app with grid layout
shinyApp(
  ui = fluidPage(
    theme = bslib::bs_theme(),
    grid_container(
      id = "main_grid",
      layout = my_layout,
      elements = list(
        header = h2(id = "header", "This is my header content"),
        sidebar = sliderInput("bins","Number of bins:", min = 1, max = 50, value = 30),
        plot = plotOutput("distPlot", height = "100\%")
      )
    )
  ),
  server = function(input, output) {
    output$distPlot <- renderPlot({
      x    <- faithful[, 2]
      bins <- seq(min(x), max(x), length.out = input$bins + 1)
      hist(x, breaks = bins, col = 'darkgray', border = 'white')
    })
  }
)
}

}
