% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_header.R
\name{grid_header}
\alias{grid_header}
\title{Header panel}
\usage{
grid_header(area, ..., tag = shiny::tags$h1)
}
\arguments{
\item{area}{Area of grid corresponding to the wrapping grid containers grid
definition}

\item{...}{Arguments passed onto the \code{tag} used. See \code{tag} argument}

\item{tag}{Tag used to build card. Defaults to \code{htmltools::div()} but can be
any function that generates a html tag as output. E.g. \code{shiny::plotOutput}.}
}
\value{

}
\description{
Makes a grid panel with no space around the top, left, and right sides.
}
\details{
To change the background color, set the css variable \code{--grid-header-bg-color}.
}
