% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_page.R
\name{grid_panel}
\alias{grid_panel}
\title{Panel element for a grid layout}
\usage{
grid_panel(id, ...)
}
\arguments{
\item{id}{Id of element in grid. This should match one of the ids in layout declaration given to \code{\link{grid_page}}.}

\item{...}{Any children elements or arguments to be passed to containing \code{\link[shiny]{div}}.}
}
\description{
Panel element for a grid layout
}
\examples{
# Only run these examples in interactive R sessions
if (interactive()) {

my_layout <- "
|      |        |         |
|------|--------|---------|
|2rem  |200px   |1fr      |
|150px |header  |header   |
|1fr   |sidebar |distPlot |"

# The classic Geyser app with grid layout
shinyApp(
  ui = grid_page(
    layout = my_layout,
    theme = bslib::bs_theme(),
    h2(id = "header", "This is my header content"),
    grid_panel(
      id = "sidebar",
      sliderInput("bins","Number of bins:", min = 1, max = 50, value = 30)
    ),
    plotOutput("distPlot", height = "100\%")
  ),
  server = function(input, output) {
    output$distPlot <- renderPlot({
      x    <- faithful[, 2]
      bins <- seq(min(x), max(x), length.out = input$bins + 1)
      hist(x, breaks = bins, col = 'darkgray', border = 'white')
    })
  }
)

}
}
\seealso{
grid_page
}
