% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_card.R
\name{grid_plot}
\alias{grid_plot}
\title{Grid Plot Output}
\usage{
grid_plot(area, outputId = area, ..., has_border = TRUE)
}
\arguments{
\item{area}{Name of grid area, should match an area defined in the layout
section of the wrapping \code{grid_page()} or \code{grid_container()}.}

\item{outputId}{Output id of the plot output. Used to link to server code
generating plot. If left unset this will use the same value as the \code{area}
argument.}

\item{...}{Elements to include within the panel}

\item{has_border}{Should the card be surrounded by a border? Set to \code{FALSE}
to turn off.}
}
\value{
A grid panel filled with plot output
}
\description{
A simple plot-on-the-grid function. For more control - such as adding headers
or footers - you should use \code{\link[=grid_card]{grid_card()}} and \code{\link[=card_plot_output]{card_plot_output()}}
directly.
}
