% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_sidebar.R
\name{grid_sidebar}
\alias{grid_sidebar}
\title{Sidebar}
\usage{
grid_sidebar(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=grid_element]{grid_element}}
  \describe{
    \item{\code{area}}{Area of grid corresponding to the wrapping grid containers grid
definition}
    \item{\code{tag}}{Tag used to build card. Defaults to \code{htmltools::div()} but can be
any function that generates a html tag as output. E.g. \code{shiny::plotOutput}.}
    \item{\code{class}}{Any additional classes you may want to add to your panel}
  }}
}
\value{

}
\description{
Makes a grid element with no space around the left, and bottom.
}
