% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridlayout-package.R
\docType{package}
\name{gridlayout-package}
\alias{gridlayout}
\alias{gridlayout-package}
\title{gridlayout: Build Layouts for R-Webapps Using CSS-Grid}
\description{
Makes it easy to layout your Shiny app or other R-derived HTML documents using CSS grid.
}
\author{
\strong{Maintainer}: Nick Strayer \email{nick.strayer@rstudio.com}

}
\keyword{internal}
