% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_gallery.R
\name{layout_gallery}
\alias{layout_gallery}
\title{Layout Gallery}
\usage{
layout_gallery()
}
\value{
A shiny app object
}
\description{
View, select, and edit layouts and generate ready-to-run app code.
}
