% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap-background-colors.R
\name{make_bg_class}
\alias{make_bg_class}
\title{Build a bootstrap background class}
\usage{
make_bg_class(bgColor = NULL, bgGradient = FALSE)
}
\arguments{
\item{bgColor}{Color of background in terms of bootstrap variable names:
Options are \code{"primary"}, \code{"secondary"}, \code{"success"}, \code{"danger"},
\code{"warning"}, \code{"info"}, \code{"light"}, and \verb{"dark}.}

\item{bgGradient}{Should a gradient be applied to the background instead of a
solid color?}
}
\value{
A string containing the bootstrap class name to apply the requested
background color
}
\description{
Info on possible colors from the \href{https://getbootstrap.com/docs/5.0/utilities/background/#background-color}{bootstrap website}
}
\examples{

gridlayout:::make_bg_class("warning")
# A gradient can be applied as well
gridlayout:::make_bg_class("primary", TRUE)

}
\keyword{internal}
