% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_to_gridlayout.R
\name{md_to_gridlayout}
\alias{md_to_gridlayout}
\title{Build grid layout obj from markdown table}
\usage{
md_to_gridlayout(
  layout_table,
  null_instead_of_error = FALSE,
  col_sizes = NULL,
  row_sizes = NULL
)
}
\arguments{
\item{layout_table}{Character string with a markdown table. First row and
column are reserved for sizing (any valid css sizing works). An optional
grid-gap can be provided using the very first cell.}

\item{null_instead_of_error}{When the input fails to be parsed as a layout,
should the function return \code{NULL} instead of throwing an error? Useful for
situations where you're parsing multiple tables simply want to check if a
table can be parsed instead of relying on it being parsable.}
}
\value{
An object of class "grid_layout", which stores the layout as a
matrix. This can be passed to other functions such as \code{layout_to_css()}.
}
\description{
Build grid layout obj from markdown table
}
\examples{
md_to_gridlayout(
  layout_table = "
    |      |120px   |1fr    |1fr    |
    |:-----|:-------|:------|:------|
    |100px |header  |header |header |
    |1fr   |sidebar |plot_a |plot_c |
    |1fr   |sidebar |plot_b |plot_b |"
)


# Can specify gap size in upper left cell
md_to_gridlayout(
  layout_table = "
    |25px  |120px  |1fr    |
    |:-----|:------|:------|
    |100px |header |header |
    |1fr   |plot   |table  |
    |1fr   |footer |footer |"
)

# Don't need to follow full md table with
# header row if not desired
md_to_gridlayout(
  layout_table = "
    |25px  |120px  |1fr    |
    |100px |header |header |
    |1fr   |plot   |table  |
    |1fr   |footer |footer |"
)

# Can omit sizing as well if desired
md_to_gridlayout(
  layout_table = "
    |header |header |
    |plot   |table  |
    |footer |footer |"
)

}
