% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_panels.R
\name{nested_grid_panel}
\alias{nested_grid_panel}
\title{Grid panel with nested gridlayout}
\usage{
nested_grid_panel(
  layout,
  title = NULL,
  elements = NULL,
  container_id = NULL,
  ...
)
}
\arguments{
\item{layout}{Either a markdown table representation (see
\code{\link{md_to_gridlayout}}) or a \code{gridlayout} object defining the
desired layout for your Shiny app.}

\item{title}{Character string to go across top of panel with label. If left
blank the card contents will take up entire space.}

\item{elements}{Named list of the UI definitions that will be used to fill
all cells. Names must match those provided in \code{layout}.}

\item{container_id}{Unique id for the layout container. This needs to be unique so
that the layout styles know where to be applied. If no value is provided
the an id is generated by prefixing with the panel and parent container ids:
e.g. \verb{#grid_page__nested_panel__grid_container}.}

\item{...}{
  Arguments passed on to \code{\link[=grid_container]{grid_container}}
  \describe{
    \item{\code{id}}{ID unique to this container (note that the HTML will be prefixed
with \verb{grid-} to avoid namespace clashes)}
    \item{\code{use_bslib_card_styles}}{Should the elements within the grid be given the
current bootstrap theme's card styling? Note that this setting will
override card styling for elements built with \link{grid_panel}.
This is so you don't have to manually change styles for each card. If you
want a mixture of card styles, then you'll need to leave this as \code{FALSE}
and set styles manually on each panel.}
    \item{\code{flag_mismatches}}{Should a mismatch between supplied \code{elements} ui
definitions and layout trigger a warning? In advanced cases you may want to
dynamically set your layout and sometimes omit panels.}
  }}
}
\value{
A \code{grid_panel} with a nested layout within it
}
\description{
Creates a panel for a layout with its own internal gridlayout
}
\seealso{
\link{grid_panel}, \link{grid_container}
}
