% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_panel.R
\name{panel}
\alias{panel}
\title{Simple grid panel}
\usage{
panel(area, ..., tag = tags$div, class = "")
}
\arguments{
\item{area}{Area of grid corresponding to the wrapping grid containers grid
definition}

\item{...}{Arguments passed onto the \code{tag} used. See \code{tag} argument}

\item{tag}{Tag used to build card. Defaults to \code{htmltools::div()} but can be
any function that generates a html tag as output. E.g. \code{shiny::plotOutput}.}

\item{class}{Any additional classes you may want to add to your panel}
}
\value{

}
\description{
The minimal adornment of a tag to be grid-aware and always able to be classed
}
