% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridded_app.R
\name{run_with_grided}
\alias{run_with_grided}
\title{Run existing app in grided layout editing mode}
\usage{
run_with_grided(app)
}
\arguments{
\item{app}{Shiny app object as produced by \code{\link[shiny]{shinyApp()}}}
}
\value{
Shiny app object with original app encapuslated by grided editor
}
\description{
Run existing app in grided layout editing mode
}
