% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinymd_to_gridlayout.R
\name{shinymd_to_gridlayout}
\alias{shinymd_to_gridlayout}
\title{Extract gridlayout from shinymd file}
\usage{
shinymd_to_gridlayout(path_to_shinymd)
}
\arguments{
\item{path_to_shinymd}{Path to a markdown file with a `.layout .grid` chunk
in it defining a grid layout}
}
\value{
Object of class `"gridlayout"`
}
\description{
Extract gridlayout from shinymd file
}
