% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_panels.R
\name{text_panel}
\alias{text_panel}
\title{Panel just containing text}
\usage{
text_panel(
  area,
  content = NULL,
  ...,
  h_align = "start",
  img_height = "55px",
  icon = NULL,
  wrapping_tag = "h2",
  is_title = FALSE
)
}
\arguments{
\item{area}{Name of grid area, should match an area defined in the layout section
of the wrapping \code{grid_page()} or \code{grid_container()}.}

\item{content}{Whatever you want the title to say. Typically just text but
any tag or tag-list is possible. All will get wrapped in an \code{h3} tag.}

\item{...}{
  Arguments passed on to \code{\link[=grid_panel]{grid_panel}}
  \describe{
    \item{\code{title}}{Character string to go across top of panel with label. If left
blank the card contents will take up entire space.}
    \item{\code{use_card_styles}}{Should the element contained in panel be made to look
like an enclosed card?}
    \item{\code{use_bslib_card_styles}}{Should the card styles from bslib be used
instead of default \code{gridlayout} card styles? If this is set to \code{TRUE} it
will override \code{use_card_style}.}
    \item{\code{collapsible}}{Should the card be able to be collapsed (\code{TRUE} or
\code{FALSE})? Gridlayout will only show collapser if the layout allows it
(panel is entirely positioned within "auto" sized rows, and has a title).
Setting this to \code{FALSE} will mean collapsibility of the panel will never be
enabled, regardless of layout.}
    \item{\code{scrollable}}{Should scroll-bars be added so content that is larger than
the panel can be seen?}
    \item{\code{v_align}}{Vertical and Horizontal alignment of content in panel.
Options include \verb{"center", "start", "end", "stretch"}. These are a direct
mapping to the the \href{https://developer.mozilla.org/en-US/docs/Web/CSS/justify-items}{css-spec for \code{justify-items}}
(=\code{h_align}) and
\href{https://developer.mozilla.org/en-US/docs/Web/CSS/align-items}{\code{align-items}}
(= \code{v_align}).}
    \item{\code{panel_id}}{ID of \code{grid_panel} div. Usually this is automatically set by
\code{grid_page()} or \code{grid_container()}.}
    \item{\code{panel_class}}{Class(es) of \code{grid_panel} div. Like \code{panel_id} this is
typically automatically generated and should only be touched in advanced
usage.}
  }}

\item{h_align}{Vertical and Horizontal alignment of content in panel.
Options include \verb{"center", "start", "end", "stretch"}. These are a direct
mapping to the the \href{https://developer.mozilla.org/en-US/docs/Web/CSS/justify-items}{css-spec for \code{justify-items}}
(=\code{h_align}) and
\href{https://developer.mozilla.org/en-US/docs/Web/CSS/align-items}{\code{align-items}}
(= \code{v_align}).}

\item{img_height}{If the passed icon is a path to an image, how tall should
that image be rendered (preserves aspect ratio.)}

\item{icon}{Optional icon/image for left of text. Supports image locations
(those ending in \verb{.png, .jpg, .jpeg} or \code{.svg}), ids of font-awesome icons
(i.e. that works with \verb{fontawesome::fa(icon)]}, or \code{fontawesome} icons as
returned by \code{\link[fontawesome:fa]{fontawesome::fa()}} (if customization of icon style is
desired.)}

\item{wrapping_tag}{What tag should the text be wrapped in. Takes either an
\code{htmltools} tag function or the string of a tag accessible via
\code{htmltools::tags[[wrapping_tag]]}.}

\item{is_title}{Should the text of this panel be passed on as the title of
the page? This will make the text show up in the browser tab or when you
bookmark the app etc..}
}
\description{
Makes a grid_panel that contains just text that is vertically centered within
the panel. Useful for titles (see \code{\link[=title_panel]{title_panel()}}) or displaying text-based
statistics.
}
\examples{

# Typically you'll just pass a character string to the function
title_panel(area = "header", "This is my header")

# Icons from `fontawesome` can be used:

# Either with just the id
text_panel(area = "header", "Here's my text", icon = "r-project")

# Or by directly passing the icon object if you want more customization
text_panel("header", "Here's my text", icon = fontawesome::fa("r-project", fill = "steelblue"))

# You can also pass arbitrary image locations for the icon
text_panel("header", "Here's my text", icon = "https://cran.r-project.org/Rlogo.svg")

# These images can have their size controlled
text_panel("header", "Here's my text", icon = "https://cran.r-project.org/Rlogo.svg", img_height = "20px")

# You can adjust the horizontal alignment of your header with h_align
text_panel("I'm in the middle", h_align = "center")

# Commonly you may want to use the text panel text as the title of your app
text_panel(area = "header", "My App Name", is_title = TRUE)

}
