% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_panel.R
\name{title_panel}
\alias{title_panel}
\title{Make header panel}
\usage{
title_panel(title_content, ..., logo = NULL)
}
\arguments{
\item{title_content}{Whatever you want the title to say. Typically just text
but any tag or tag-list is possible. All will get wrapped in an \code{h3} tag.}

\item{...}{
  Arguments passed on to \code{\link[=grid_panel]{grid_panel}}
  \describe{
    \item{\code{use_bslib_card_styles}}{Should the card styles from bslib be used
instead of default \code{gridlayout} card styles? If this is set to \code{TRUE} it
will override \code{use_card_style}.}
    \item{\code{collapsable}}{Should the card be able to be collapsed? Only really
functional when panel takes up entire width of page. Requires a title so
the user knows what panel is in collapsed state.}
    \item{\code{scrollable}}{Should scroll-bars be added so content that is larger than
the panel can be seen?}
    \item{\code{v_align}}{Vertical and Horizontal alignment of content in panel.
Options include \verb{"center", "start", "end", "stretch"}. These are a direct
mapping to the the \href{https://developer.mozilla.org/en-US/docs/Web/CSS/justify-items}{css-spec for \code{justify-items}}
(=\code{h_align}) and
\href{https://developer.mozilla.org/en-US/docs/Web/CSS/align-items}{\code{align-items}}
(= \code{v_align}).}
    \item{\code{h_align}}{Vertical and Horizontal alignment of content in panel.
Options include \verb{"center", "start", "end", "stretch"}. These are a direct
mapping to the the \href{https://developer.mozilla.org/en-US/docs/Web/CSS/justify-items}{css-spec for \code{justify-items}}
(=\code{h_align}) and
\href{https://developer.mozilla.org/en-US/docs/Web/CSS/align-items}{\code{align-items}}
(= \code{v_align}).}
  }}

\item{logo}{Optional image source for logo to be placed left of title}
}
\description{
This is just a helper function that wraps your content in a vertically
centered header (\code{h2}) tag.
}
\examples{

# Typically you'll just pass a character string to the function
title_panel("This is my header")

# You can adjust the horizontal alignment of your header with h_align
title_panel("I'm on the right side!", h_align = "end")

# If you want more than just text in your title, just wrap content in a span tag
requireNamespace("fontawesome", quietly = TRUE)
title_panel(
  htmltools::span(
    fontawesome::fa("r-project", fill = "steelblue"),
    "This is my header"
  )
)

}
