% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_app_template.R
\name{to_app_template}
\alias{to_app_template}
\title{Build Shiny app template based on a grid layout}
\usage{
to_app_template(layout)
}
\arguments{
\item{layout}{Object of class \code{"gridlayout"}.}
}
\value{
Character string of Shiny app that uses desired layout
}
\description{
Returns the string of an entire shiny app built around a desired layout.
Designed to be used with the \code{gridlayout} app-starter addin.
}
\examples{

}
