% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_md.R
\name{to_matrix}
\alias{to_matrix}
\title{Convert \code{gridlayout} to matrix format}
\usage{
to_matrix(layout)
}
\arguments{
\item{layout}{A gridlayout object (as created by \code{new_gridlayout()})}
}
\value{
A matrix of the layout with each grid cell and its item membership
corresponding to a cell in the matrix
}
\description{
Convert \code{gridlayout} to matrix format
}
\examples{

layout <- md_to_gridlayout("
|10px  |120px   |1fr    |1fr    |
  |100px |header  |header |header |
  |1fr   |sidebar |plot_a |plot_c |
  |1fr   |sidebar |plot_b |plot_b |")

to_matrix(layout)

}
