% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_md.R
\name{to_md}
\alias{to_md}
\title{Convert layout to markdown table}
\usage{
to_md(layout)
}
\arguments{
\item{layout}{Object of class `"gridlayout"`.}
}
\value{
Markdown table as constructed by `knitr::kable()` that defines the
  grid layout. This can be used with the function `grid_layout_from_md()` to
  build grid layouts.
}
\description{
Uses the `knitr::kable()` function to convert a layout object to a markdown
table
}
\examples{

start_table <- "
|      |        |       |       |
|:-----|:-------|:------|:------|
|10px  |120px   |1fr    |1fr    |
|100px |header  |header |header |
|1fr   |sidebar |plot_a |plot_c |
|1fr   |sidebar |plot_b |plot_b |"

my_layout <- grid_layout_from_md(layout_table = start_table)
to_md(my_layout)

}
