% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_to_gridlayout.R
\name{update_rmd}
\alias{update_rmd}
\title{Update a layout chunk in a rmd document}
\usage{
update_rmd(path_to_rmd, new_layout, updated_rmd_path = path_to_rmd)
}
\arguments{
\item{path_to_rmd}{Path to a markdown file with a \verb{.layout .grid} chunk
in it defining a grid layout}

\item{new_layout}{An updated \code{gridlayout} object to replace the current one}

\item{updated_rmd_path}{Where the rmd with the new layout should be
placed. Default is overwriting the previous file}
}
\description{
Update a layout chunk in a rmd document
}
\examples{

loc_of_rmd <- system.file("sample_apps/my_app.Rmd", package = "gridlayout")
first_layout <- rmd_to_gridlayout(loc_of_rmd)
first_layout

new_layout <- md_to_gridlayout(
  layout_table = "
    |      |120px   |1fr    |1fr    |
    |:-----|:-------|:------|:------|
    |100px |header  |header |header |
    |1fr   |sidebar |plot_a |plot_c |
    |1fr   |sidebar |plot_b |plot_b |"
)

loc_of_new_rmd <- tempfile(fileext = ".rmd")
update_rmd(loc_of_rmd, new_layout, loc_of_new_rmd)
rmd_to_gridlayout(loc_of_new_rmd)

}
