% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinymd_to_gridlayout.R
\name{update_shinymd}
\alias{update_shinymd}
\title{Update a layout chunk in a shinymd document}
\usage{
update_shinymd(
  path_to_shinymd,
  new_layout,
  updated_shinymd_path = path_to_shinymd
)
}
\arguments{
\item{path_to_shinymd}{Path to a markdown file with a `.layout .grid` chunk
in it defining a grid layout}

\item{new_layout}{An updated `gridlayout` object to replace the current one}

\item{updated_shinymd_path}{Where the shinymd with the new layout should be
placed. Default is overwriting the previous file}
}
\description{
Update a layout chunk in a shinymd document
}
\examples{

loc_of_shinymd <- system.file("sample_apps/my_app.Rmd", package = "gridlayout")
first_layout <- shinymd_to_gridlayout(loc_of_shinymd)
first_layout

new_layout <- md_to_gridlayout(
  layout_table = "
    |      |120px   |1fr    |1fr    |
    |:-----|:-------|:------|:------|
    |100px |header  |header |header |
    |1fr   |sidebar |plot_a |plot_c |
    |1fr   |sidebar |plot_b |plot_b |"
)

loc_of_new_shinymd <- tempfile(fileext = ".rmd")
update_shinymd(loc_of_shinymd, new_layout, loc_of_new_shinymd)
shinymd_to_gridlayout(loc_of_new_shinymd)

}
