% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_css_rule.R
\name{build_css_rule}
\alias{build_css_rule}
\title{Build css properties named or unnamed list of property values}
\usage{
build_css_rule(selector, prop_list)
}
\arguments{
\item{selector}{valid css selector to target. E.g. \code{body} or \code{div.blue_boxes}...}

\item{prop_list}{A list of property-value pairs for additional styles to
be added to each element. Pairs can be given as named elements: e.g.
\code{element_styles = c("background" = "blue")}, or simple as the whole
character: e.g. \code{element_styles = c("background: blue")}. (Semicolon usage
doesn't matter for single character declaration.)}
}
\value{
A concatenated string of property values to be used inside a css selector.
}
\description{
Build css properties named or unnamed list of property values
}
\examples{

custom_styles <- c(
  "background" = "blue",
  "border" = "1px solid red",
  "cursor: none"
)
build_css_rule(custom_styles)

}
