% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex_stack.R
\name{flex_stack}
\alias{flex_stack}
\title{Grid panel with vertically stacked elements}
\usage{
flex_stack(
  area,
  ...,
  item_alignment = "top",
  title = NULL,
  collapsible = TRUE,
  scrollable = FALSE,
  item_gap = "10px"
)
}
\arguments{
\item{area}{Name of grid area, should match an area defined in the layout
section of the wrapping \code{grid_page()} or \code{grid_container()}.}

\item{...}{Elements to include within the panel}

\item{item_alignment}{How should the items within the panel be aligned.
Defaults to stacking downward from the top. Options include \code{"top"},
\code{"centered"}, \code{"bottom"}, or \code{"spread"} (items are distributed evenly among
vertical space).}

\item{title}{Character string to go across top of panel with label. If left
blank the card contents will take up entire space.}

\item{collapsible}{Should the card be able to be collapsed (\code{TRUE} or
\code{FALSE})? Gridlayout will only show collapser if the layout allows it
(panel is entirely positioned within "auto" sized rows, and has a title).
Setting this to \code{FALSE} will mean collapsibility of the panel will never be
enabled, regardless of layout.}

\item{scrollable}{Should scroll-bars be added so content that is larger than
the panel can be seen?}

\item{item_gap}{How much space should there be between consecutive items?}

\item{use_card_styles}{Should the element contained in panel be made to look
like an enclosed card?}

\item{v_align, h_align}{Vertical and Horizontal alignment of content in panel.
Options include \verb{"center", "start", "end", "stretch"}. These are a direct
mapping to the the \href{https://developer.mozilla.org/en-US/docs/Web/CSS/justify-items}{css-spec for \code{justify-items}}
(=\code{h_align}) and
\href{https://developer.mozilla.org/en-US/docs/Web/CSS/align-items}{\code{align-items}}
(= \code{v_align}).}

\item{panel_id}{ID of \code{grid_card} div. Usually this is automatically set by
\code{grid_page()} or \code{grid_container()}.}

\item{panel_class}{Class(es) of \code{grid_card} div. Like \code{panel_id} this is
typically automatically generated and should only be touched in advanced
usage.}

\item{padding}{A valid css size that overrides the default of how much
spacing is around the card contents}
}
\value{
Elements from \code{...} wrapped in a \code{shiny::div()} with styles for
vertical stacking applied.
}
\description{
Contain a series of ui elements with each vertically stacked on top of
eachother. Internally uses css flexbox to align items.
}
\examples{

# Simply a wrapper for shiny::div(...)
flex_stack(
  area="header",
  shiny::h2("R"),
  shiny::actionButton("myButton", "Click Me")
)

}
\seealso{
\link{grid_card}
}
