% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elements.R
\name{get_elements}
\alias{get_elements}
\title{Get elements and positions out of a grid layout}
\usage{
get_elements(layout)
}
\arguments{
\item{layout}{Object of class \code{"gridlayout"}.}
}
\value{
List of all unique elements in the layout and their \code{id},
\verb{\{start,end\}_row}, and \verb{\{start,end\}_col}. Positions are indexed starting at
1
}
\description{
Get elements and positions out of a grid layout
}
\examples{
grid_obj <- md_to_gridlayout(
  layout_table = "
        |      |120px   |1fr    |1fr    |
        |------|--------|-------|-------|
        |100px |header  |header |header |
        |1fr   |sidebar |plot_a |plot_a |
        |1fr   |sidebar |plot_b |plot_c |"
 )

 get_elements(grid_obj)

}
\seealso{
\link{get_element_ids}
}
