% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_card.R
\name{grid_card}
\alias{grid_card}
\title{Grid Card}
\usage{
grid_card(
  area,
  ...,
  title = NULL,
  scrollable = FALSE,
  collapsible = TRUE,
  has_border = TRUE,
  item_gap = "10px"
)
}
\arguments{
\item{area}{Name of grid area, should match an area defined in the layout
section of the wrapping \code{grid_page()} or \code{grid_container()}.}

\item{...}{Elements to include within the panel}

\item{title}{Character string to go across top of panel with label. If left
blank the card contents will take up entire space.}

\item{scrollable}{Should scroll-bars be added so content that is larger than
the panel can be seen?}

\item{collapsible}{Should the card be able to be collapsed (\code{TRUE} or
\code{FALSE})? Gridlayout will only show collapser if the layout allows it
(panel is entirely positioned within "auto" sized rows, and has a title).
Setting this to \code{FALSE} will mean collapsibility of the panel will never be
enabled, regardless of layout.}

\item{has_border}{Should the card be surrounded by a border? Set to \code{FALSE}
to turn off.}

\item{item_gap}{How much space should there be between consecutive items?}
}
\value{

}
\description{
The standard element for placing elements on the grid in a simple card
container
}
\seealso{
\link{grid_card_text}
}
