% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_panel.R
\name{grid_panel}
\alias{grid_panel}
\title{Panel for grid page}
\usage{
grid_panel(..., v_align, h_align)
}
\arguments{
\item{...}{Elements to include within the panel}

\item{v_align, h_align}{Vertical and Horizontal alignment of content in panel.
Options include \verb{"center", "start", "end", "stretch"}. These are a direct
mapping to the the \href{https://developer.mozilla.org/en-US/docs/Web/CSS/justify-items}{css-spec for \code{justify-items}}
(=\code{h_align}) and
\href{https://developer.mozilla.org/en-US/docs/Web/CSS/align-items}{\code{align-items}}
(= \code{v_align}).}
}
\value{
Elements from \code{...} wrapped in a \code{shiny::div()} with \code{class = "grid_panel"} and vertical/horizontal alignment css applied.
}
\description{
Panel for grid page
}
