% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_element.R
\name{grid_place}
\alias{grid_place}
\title{Place an item on the grid}
\usage{
grid_place(area, element)
}
\arguments{
\item{area}{Area of grid corresponding to the wrapping grid containers grid
definition}

\item{element}{Element (html tag) to be placed.}
}
\value{
The \code{element} with a \code{grid-area} property added to its styles
}
\description{
Place an item on the grid
}
\examples{

grid_place("header", htmltools::tags$h1("I am a header"))

}
