% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_plot.R
\name{grid_plot}
\alias{grid_plot}
\title{Grid Plot Output}
\usage{
grid_plot(area, outputId = area, ...)
}
\arguments{
\item{area}{Name of grid area, should match an area defined in the layout
section of the wrapping \code{grid_page()} or \code{grid_container()}.}

\item{outputId}{Output id of the plot output. Used to link to server code
generating plot. If left unset this will use the same value as the \code{area}
argument.}

\item{...}{
  Arguments passed on to \code{\link[bslib:card_body]{bslib::card_plot_output}}
  \describe{
    \item{\code{stretch}}{Set to \code{TRUE} if this \code{card_body} is eager to use any extra
vertical space is available in the card.}
  }}
}
\value{
A grid panel filled with plot output
}
\description{
A simple plot-on-the-grid function. For more control - such as adding headers
or footers - you should use \code{\link[=grid_card]{grid_card()}} and \code{\link[bslib:card_body]{bslib::card_plot_output()}}
directly.
}
