% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_to_gridlayout.R
\name{md_to_gridlayout}
\alias{md_to_gridlayout}
\title{Build grid layout obj from markdown table}
\usage{
md_to_gridlayout(layout_table)
}
\arguments{
\item{layout_table}{Character string with a markdown table. First row and
column are reserved for sizing (any valid css sizing works). An optional
grid-gap can be provided using the very first cell.}
}
\value{
An object of class "grid_layout", which stores the layout as a
  matrix. This can be passed to other functions such as `layout_to_css()`.
}
\description{
Build grid layout obj from markdown table
}
\examples{
grid_obj <- md_to_gridlayout(
  layout_table = "
    |      |120px   |1fr    |1fr    |
    |:-----|:-------|:------|:------|
    |100px |header  |header |header |
    |1fr   |sidebar |plot_a |plot_c |
    |1fr   |sidebar |plot_b |plot_b |"
)

}
