% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_panels.R
\name{nested_grid_panel}
\alias{nested_grid_panel}
\title{Grid panel with nested gridlayout}
\usage{
nested_grid_panel(
  layout,
  title = NULL,
  elements = NULL,
  container_id = NULL,
  use_bslib_card_styles = FALSE,
  flag_mismatches = TRUE
)
}
\arguments{
\item{layout}{Either a markdown table representation (see
\code{\link{md_to_gridlayout}}) or a \code{gridlayout} object defining the
desired layout for your Shiny app.}

\item{title}{Character string to go across top of panel with label. If left
blank the card contents will take up entire space.}

\item{container_id}{Unique id for the layout container. This needs to be unique so
that the layout styles know where to be applied. If no value is provided
the an id is generated by prefixing with the panel and parent container ids:
e.g. \verb{#grid_page__nested_panel__grid_container}.}

\item{use_bslib_card_styles}{Should the card styles from bslib be used
instead of default \code{gridlayout} card styles? If this is set to \code{TRUE} it
will override \code{use_card_style}.}

\item{flag_mismatches}{Should a mismatch between supplied \code{elements} ui
definitions and layout trigger a warning? In advanced cases you may want to
dynamically set your layout and sometimes omit panels.}
}
\value{
A \code{grid_panel} with a nested layout within it
}
\description{
Creates a panel for a layout with its own internal gridlayout
}
\seealso{
\link{grid_panel}, \link{grid_container}
}
