% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_grid_panel.R
\name{nested_grid_panel}
\alias{nested_grid_panel}
\title{Grid panel with nested gridlayout}
\usage{
nested_grid_panel(
  area,
  layout,
  ...,
  id = NULL,
  title = NULL,
  use_bslib_card_styles = FALSE,
  flag_mismatches = TRUE
)
}
\arguments{
\item{area}{Name of grid area, should match an area defined in the layout section
of the wrapping \code{grid_page()} or \code{grid_container()}.}

\item{layout}{Either a markdown table representation (see
\code{\link{md_to_gridlayout}}) or a \code{gridlayout} object defining the
desired layout for your Shiny app.}

\item{...}{Elements to include within the panel}

\item{id}{ID unique to this container (note that the HTML will be prefixed
with \verb{grid-} to avoid namespace clashes)}

\item{title}{Character string to go across top of panel with label. If left
blank the card contents will take up entire space.}

\item{use_bslib_card_styles}{Should the card styles from bslib be used
instead of default \code{gridlayout} card styles? If this is set to \code{TRUE} it
will override \code{use_card_style}.}

\item{flag_mismatches}{Should a mismatch between supplied \code{elements} ui
definitions and layout trigger a warning? In advanced cases you may want to
dynamically set your layout and sometimes omit panels.}
}
\value{
A \code{grid_panel} with a nested layout within it
}
\description{
Creates a panel for a layout with its own internal gridlayout
}
\examples{
if (interactive()) {
library(gridlayout)
library(shiny)

# A centered text panel
text_panel_c <- function(...) text_panel(..., h_align="center")

my_layout <- "
|A |B |
|C |D |"
shinyApp(
  ui = grid_page(
    layout = my_layout,
    text_panel_c("A","A"),
    text_panel_c("B","B"),
    text_panel_c("C","C"),
    nested_grid_panel(
      "D",
      layout = my_layout,
      text_panel_c("A","A2"),
      text_panel_c("B","B2"),
      text_panel_c("C","C2"),
      text_panel_c("D","D2")
    )
  ),
  server = function(input, output) {}
)
}
}
\seealso{
\link{grid_panel}, \link{grid_container}
}
