% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_gridlayout.R
\name{new_gridlayout}
\alias{new_gridlayout}
\title{Construct a gridlayout object from basic parts}
\usage{
new_gridlayout(layout_mat, col_sizes, row_sizes, gap, element_list)
}
\arguments{
\item{layout_mat}{A character matrix where each entry controls what element
id will take up that given position in the grid. For instance a header
element would take up every cell in the first row.}

\item{col_sizes}{A character vector of valid css sizes for the width of each
column in your grid as given by \code{layout_mat}. If a single value is passed, it will be repeated for all columns.}

\item{row_sizes}{Same as \code{col_sizes}, but for row heights.}

\item{gap}{Valid css sizing for gap to be left between each element in your
grid. Defaults to \code{"1rem"}.}
}
\value{
Object of class \code{"gridlayout"}
}
\description{
Construct a gridlayout object from basic parts
}
\examples{

my_layout <- new_gridlayout(
  layout_mat = matrix(c(
    "header", "header",
    "plota",  "plotb"
  ), ncol = 2, byrow = TRUE),
  col_sizes = c("1fr", "2fr"),
  row_sizes = c("100px", "1fr"),
  gap = "2rem"
)

}
