% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_to_gridlayout.R
\name{rmd_to_gridlayout}
\alias{rmd_to_gridlayout}
\title{Extract gridlayout from rmd file}
\usage{
rmd_to_gridlayout(path_to_rmd)
}
\arguments{
\item{path_to_rmd}{Path to a markdown file with a \verb{.layout .grid} chunk
in it defining a grid layout}
}
\value{
Object of class \code{"gridlayout"}
}
\description{
Extract gridlayout from rmd file
}
\examples{

my_app_loc <- system.file("sample_apps/my_app.Rmd", package = "gridlayout")
my_layout <- rmd_to_gridlayout(my_app_loc)
my_layout

}
