% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinymd_to_gridlayout.R
\name{shinymd_to_gridlayout}
\alias{shinymd_to_gridlayout}
\title{Extract gridlayout from shinymd file}
\usage{
shinymd_to_gridlayout(path_to_shinymd)
}
\arguments{
\item{path_to_shinymd}{Path to a markdown file with a \verb{.layout .grid} chunk
in it defining a grid layout}
}
\value{
Object of class \code{"gridlayout"}
}
\description{
Extract gridlayout from shinymd file
}
\examples{

my_app_loc <- system.file("sample_apps/my_app.Rmd", package = "gridlayout")
my_layout <- shinymd_to_gridlayout(my_app_loc)
my_layout

}
