% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_panels.R
\name{title_panel}
\alias{title_panel}
\title{Make header panel}
\usage{
title_panel(
  area,
  title,
  windowTitle = title,
  h_align = "start",
  img_height = "55px",
  icon = NULL
)
}
\arguments{
\item{area}{Name of grid area, should match an area defined in the layout
section of the wrapping \code{grid_page()} or \code{grid_container()}.}

\item{title}{An application title to display.}

\item{windowTitle}{The title that should be displayed by the browser window.}

\item{h_align}{Vertical and Horizontal alignment of content in panel.
Options include \verb{"center", "start", "end", "stretch"}. These are a direct
mapping to the the \href{https://developer.mozilla.org/en-US/docs/Web/CSS/justify-items}{css-spec for \code{justify-items}}
(=\code{h_align}) and
\href{https://developer.mozilla.org/en-US/docs/Web/CSS/align-items}{\code{align-items}}
(= \code{v_align}).}

\item{img_height}{If the passed icon is a path to an image, how tall should
that image be rendered (preserves aspect ratio.)}

\item{icon}{Optional icon/image for left of text. Supports image locations
(those ending in \verb{.png, .jpg, .jpeg} or \code{.svg}), ids of font-awesome icons
(i.e. that works with \verb{fontawesome::fa(icon)]}, or \code{fontawesome} icons as
returned by \code{\link[fontawesome:fa]{fontawesome::fa()}} (if customization of icon style is
desired.)}
}
\description{
This is just a helper function that wraps your content in a vertically
centered header (\code{h2}) tag via the \code{\link[=text_panel]{text_panel()}} function. Also adds the
apps content to the \code{title} metatag of your webpage so it shows up properly
in tabs etc.. Modeled after \code{\link[shiny:titlePanel]{shiny::titlePanel()}}.
}
\examples{

# Simple app title
title_panel(area = "header", "My App")

# Centered title
title_panel("header", "My App (in the center)", h_align = "center")

# Can have a logo
title_panel("footer", "My App", icon = "https://cran.r-project.org/Rlogo.svg")

}
\seealso{
\link{text_panel}
}
