% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_css.R
\name{to_css}
\alias{to_css}
\title{Convert grid layout to css string}
\usage{
to_css(
  layout,
  container,
  use_card_style = TRUE,
  element_styles = c(),
  debug_mode = FALSE,
  full_height = TRUE
)
}
\arguments{
\item{layout}{Object of class \code{"gridlayout"}.}

\item{container}{Id of the element for grid to be placed in. Default value to
apply grid styling to the whole app (aka the \code{body} element) for whole page
grids}

\item{use_card_style}{Should each section of the grid be given a card style
to make it stand out?}

\item{element_styles}{A list of named property-value pairs for additional
styles to be added to each element. E.g. \code{element_styles = c("background" = "blue")}.}

\item{debug_mode}{If set to \code{TRUE} then each element of the grid will have an
outline applied so positioning can more easily be assessed.}

\item{full_height}{Should the grid-containing element be made as tall as
possible? Set to \code{FALSE} if the grid is contained within another element
and is not the whole app's UI.}
}
\value{
Character string of css used to setup grid layout and place elements
(referenced by id) into correct locations
}
\description{
Convert grid layout to css string
}
\examples{

grid_obj <- md_to_gridlayout(
  layout_table = "
    |      |120px   |1fr    |1fr    |
    |:-----|:-------|:------|:------|
    |100px |header  |header |header |
    |1fr   |sidebar |plot_a |plot_c |
    |1fr   |sidebar |plot_b |plot_b |"
)

to_css(grid_obj)

}
