% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_css.R
\name{to_css}
\alias{to_css}
\title{Convert grid layout to css string}
\usage{
to_css(layout, container, debug_mode = FALSE)
}
\arguments{
\item{layout}{Object of class `"gridlayout"`.}

\item{container}{Id of the element for grid to be placed in. Default value to
apply grid styling to the whole app (aka the `body` element) for whole page
grids}

\item{debug_mode}{If set to `TRUE` then each element of the grid will have an
outline applied so positioning can more easily be assessed.}
}
\value{
Character string of css used to setup grid layout and place elements
  (referenced by id) into correct locations
}
\description{
Convert grid layout to css string
}
\examples{

grid_obj <- md_to_gridlayout(
  layout_table = "
    |      |120px   |1fr    |1fr    |
    |:-----|:-------|:------|:------|
    |100px |header  |header |header |
    |1fr   |sidebar |plot_a |plot_c |
    |1fr   |sidebar |plot_b |plot_b |"
)

to_css(grid_obj)

}
