% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_css.R
\name{to_css}
\alias{to_css}
\title{Convert grid layout to css string}
\usage{
to_css(layout, ...)
}
\arguments{
\item{layout}{Object of class \code{"gridlayout"}.}

\item{container}{Id of the element for grid to be placed in. Default value to
apply grid styling to the whole app (aka the \code{body} element) for whole page
grids. If an simple name is passed it is assumed that it is an id: i.e. the
css id selector \verb{#} is apended to the name. If container contains css
selector characters such as a dot, the selector will not be transformed
into an id automatically. E.g. \code{container = ".main-content"}.}

\item{is_card_styled}{Should each section of the grid be given a card style
to make it stand out? Options are \code{"grid_panel"}, where only elements with
\code{"grid_panel"} class will get card styling, \code{"all"} where all children of
the grid container will get card styling (useful for RMarkdown or other
situations where you don't control child rendering) or \code{"none"} for no card
styling.}

\item{additional_card_styles}{A list of named property-value pairs for
additional styles to be added to each card-element (as defined by
\code{is_card_styled}. E.g. \code{element_styles = c("background" = "blue")}.}

\item{element_styles}{A list of named property-value pairs for additional
styles to be added to each first child element of the grid container. E.g.
under the css selector of \verb{grid_container > *}. If \code{is_card_styled = "all"}
then this has the same purpose as as \code{additional_card_styles}}

\item{debug_mode}{If set to \code{TRUE} then each element of the grid will have an
outline applied so positioning can more easily be assessed.}

\item{container_height}{How tall should the grid-containing element be? If
set to \code{"viewport"} then the container will take up the entire available
vertical space (equivalent to the CSS value of \verb{100vh}). For nested grids a
value of \code{"100\%"} will enable the the nested grid to take up the full
vertical space provided by its parent container. Any other css size unit is
also valid such as \code{"400px"} or \code{"100\%"}.}

\item{selector_prefix}{CSS prefix used to target grid elements. This will
change if you're integrating grid with a system that you don't want to use
ids (the \code{"#"} prefix) with because they are not available or are used for
other reasons.}
}
\value{
Character string of css used to setup grid layout and place elements
(referenced by id) into correct locations
}
\description{
Convert grid layout to css string
}
\examples{

grid_obj <- md_to_gridlayout(
  layout_table = "
    |      |120px   |1fr    |1fr    |
    |:-----|:-------|:------|:------|
    |100px |header  |header |header |
    |1fr   |sidebar |plot_a |plot_c |
    |1fr   |sidebar |plot_b |plot_b |"
)

cat(to_css(grid_obj))

}
