% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_css.R
\name{to_css}
\alias{to_css}
\title{Convert grid layout to css string}
\usage{
to_css(
  layout,
  container,
  is_card_styled = "grid_panel",
  element_styles = c(),
  selector_prefix = "#"
)
}
\arguments{
\item{layout}{Object of class \code{"gridlayout"}.}

\item{container}{Id of the element for grid to be placed in. Default value to
apply grid styling to the whole app (aka the \code{body} element) for whole page
grids. If an simple name is passed it is assumed that it is an id: i.e. the
css id selector \verb{#} is apended to the name. If container contains css
selector characters such as a dot, the selector will not be transformed
into an id automatically. E.g. \code{container = ".main-content"}.}

\item{is_card_styled}{Should each section of the grid be given a card style
to make it stand out? Options are \code{"grid_panel"}, where only elements with
\code{"grid_panel"} class will get card styling, \code{"all"} where all children of
the grid container will get card styling (useful for RMarkdown or other
situations where you don't control child rendering) or \code{"none"} for no card
styling.}

\item{element_styles}{A list of named property-value pairs ( E.g.
\code{element_styles = c("background" = "blue")}.) for additional styles to be
added to each first child element of the grid container. E.g. under the css
selector of \verb{grid_container > *}. If \code{is_card_styled = "all"} then this has
the same purpose as as \code{additional_card_styles}}

\item{selector_prefix}{CSS prefix used to target grid elements. This will
change if you're integrating grid with a system that you don't want to use
ids (the \code{"#"} prefix) with because they are not available or are used for
other reasons.}
}
\value{
Character string of css used to setup grid layout and place elements
(referenced by id) into correct locations
}
\description{
Convert grid layout to css string
}
\examples{

grid_obj <- md_to_gridlayout(
  layout_table = "
    |      |120px   |1fr    |1fr    |
    |:-----|:-------|:------|:------|
    |100px |header  |header |header |
    |1fr   |sidebar |plot_a |plot_c |
    |1fr   |sidebar |plot_b |plot_b |"
)

cat(to_css(grid_obj))

}
