% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_md.R
\name{to_md}
\alias{to_md}
\title{Convert layout to markdown table}
\usage{
to_md(layout, include_gap_size = TRUE)
}
\arguments{
\item{layout}{Object of class \code{"gridlayout"}.}

\item{include_gap_size}{Should the gap size for the layout be added in
upper-left of table?}
}
\value{
Markdown table that defines the grid layout. This can be used with
the function \code{grid_layout_from_md()} to build grid layouts.
}
\description{
Convert layout to markdown table
}
\examples{

start_table <- "
|      |        |       |       |
|:-----|:-------|:------|:------|
|10px  |120px   |1fr    |1fr    |
|100px |header  |header |header |
|1fr   |sidebar |plot_a |plot_c |
|1fr   |sidebar |plot_b |plot_b |"

my_layout <- md_to_gridlayout(layout_table = start_table)
table_text <- to_md(my_layout)

}
