% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_css.R
\name{use_gridlayout}
\alias{use_gridlayout}
\title{Convert layout to css for Shiny}
\usage{
use_gridlayout(layout, ...)
}
\arguments{
\item{layout}{Either a markdown table representation (see
\code{\link{md_to_gridlayout}}) or a \code{gridlayout} object defining the
desired layout for your Shiny app.}

\item{...}{
  Arguments passed on to \code{\link[=to_css]{to_css}}
  \describe{
    \item{\code{container}}{Id of the element for grid to be placed in. Default value to
apply grid styling to the whole app (aka the \code{body} element) for whole page
grids. If an simple name is passed it is assumed that it is an id: i.e. the
css id selector \verb{#} is apended to the name. If container contains css
selector characters such as a dot, the selector will not be transformed
into an id automatically. E.g. \code{container = ".main-content"}.}
    \item{\code{use_card_style}}{Should each section of the grid be given a card style
to make it stand out?}
    \item{\code{element_styles}}{A list of named property-value pairs for additional
styles to be added to each element. E.g. \code{element_styles = c("background" = "blue")}.}
    \item{\code{debug_mode}}{If set to \code{TRUE} then each element of the grid will have an
outline applied so positioning can more easily be assessed.}
    \item{\code{full_height}}{Should the grid-containing element be made as tall as
possible? Set to \code{FALSE} if the grid is contained within another element
and is not the whole application's UI. In this case the grid-container will take
the full height of containing element.}
    \item{\code{selector_prefix}}{CSS prefix used to target grid elements. This will
change if you're integrating grid with a system that you don't want to use
ids (the \code{"#"} prefix) with because they are not available or are used for
other reasons.}
  }}
}
\value{
Character string of css used to setup grid layout and place elements
(referenced by id) into correct locations
}
\description{
This simply wraps the output of \code{to_css()} in a \code{style} tag and escapes HTML
characters to simplify using in Shiny
}
\examples{
# Only run these examples in interactive R sessions
if (interactive()) {

my_layout <- "
|      |        |         |
|------|--------|---------|
|2rem  |200px   |1fr      |
|100px |header  |header   |
|1fr   |sidebar |distPlot |"

# The classic Geyser app with grid layout
shinyApp(
  ui = fluidPage(
    use_gridlayout(my_layout, "app-container"),
    div(
      id = "app-container",
      div(id = "header",
          h2(id = "app-title", "Old Faithful Geyser Data")
      ),
      div(id = "sidebar",
          sliderInput("bins","Number of bins:",
                      min = 1, max = 50, value = 30)
      ),
      plotOutput("distPlot")
    )
  ),
  server = function(input, output) {
    output$distPlot <- renderPlot({
      x    <- faithful[, 2]
      bins <- seq(min(x), max(x), length.out = input$bins + 1)
      hist(x, breaks = bins, col = 'darkgray', border = 'white')
    })
  }
)
}
}
