% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_css.R
\name{use_gridlayout_rmd}
\alias{use_gridlayout_rmd}
\title{Enable gridlayout usage in RMarkdown documents}
\usage{
use_gridlayout_rmd()
}
\description{
Adds required hooks to RMarkdown to process \code{gridlayout} chunks and style document accordingly.
}
