% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_gridlayout_rmd.R
\name{use_gridlayout_rmd}
\alias{use_gridlayout_rmd}
\title{Enable \code{gridlayout} usage in RMarkdown documents}
\usage{
use_gridlayout_rmd(container_query = ".main-container", selector_prefix = "#")
}
\arguments{
\item{container_query}{The CSS selector used to access the main grid
container. This is typically left at the default of \code{'.main-container'} as
that's the parent of each section in the standard rmd template.}

\item{selector_prefix}{CSS prefix used to target grid elements. This will
change if you're integrating grid with a system that you don't want to use
ids (the \code{"#"} prefix) with because they are not available or are used for
other reasons.}
}
\description{
Adds required hooks to RMarkdown to process \code{gridlayout} chunks and style
document accordingly. Layout will be generated from a chunk identified with
the syntax \verb{```\{gridlayout\} <insert layout table here> ```}
}
\details{
See \code{vignette("using_with_rmd", package = "gridlayout")} for a more in-depth
overview.
}
