% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_css.R
\name{use_gridlayout_rmd}
\alias{use_gridlayout_rmd}
\title{Enable gridlayout usage in RMarkdown documents}
\usage{
use_gridlayout_rmd(
  container = ".main-container",
  is_card_styled = "all",
  element_styles = NULL,
  selector_prefix = "#"
)
}
\arguments{
\item{container}{Id of the element for grid to be placed in. Default value to
apply grid styling to the whole app (aka the \code{body} element) for whole page
grids. If an simple name is passed it is assumed that it is an id: i.e. the
css id selector \verb{#} is apended to the name. If container contains css
selector characters such as a dot, the selector will not be transformed
into an id automatically. E.g. \code{container = ".main-content"}.}

\item{is_card_styled}{Should each section of the grid be given a card style
to make it stand out? Options are \code{"grid_panel"}, where only elements with
\code{"grid_panel"} class will get card styling, \code{"all"} where all children of
the grid container will get card styling (useful for RMarkdown or other
situations where you don't control child rendering) or \code{"none"} for no card
styling.}

\item{element_styles}{A list of named property-value pairs ( E.g.
\code{element_styles = c("background" = "blue")}.) for additional styles to be
added to each first child element of the grid container. E.g. under the css
selector of \verb{grid_container > *}. If \code{is_card_styled = "all"} then this has
the same purpose as as \code{additional_card_styles}}

\item{selector_prefix}{CSS prefix used to target grid elements. This will
change if you're integrating grid with a system that you don't want to use
ids (the \code{"#"} prefix) with because they are not available or are used for
other reasons.}
}
\description{
Adds required hooks to RMarkdown to process \code{gridlayout} chunks and style
document accordingly. Layout will be generated from a chunk identified with
the syntax\preformatted{```\{gridlayout\}
<insert layout table here>
```
}
}
\details{
See \code{vignette("using_with_rmd", package = "gridlayout")} for a more in-depth overview.
}
