% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_css.R
\name{use_gridlayout_rmd}
\alias{use_gridlayout_rmd}
\title{Enable gridlayout usage in RMarkdown documents}
\usage{
use_gridlayout_rmd(
  container = ".main-container",
  is_card_styled = "all",
  selector_prefix = "#"
)
}
\arguments{
\item{is_card_styled}{Should each section of the grid be given a card style
to make it stand out? Options are \code{"grid_panel"}, where only elements with
\code{"grid_panel"} class will get card styling, \code{"all"} where all children of
the grid container will get card styling (useful for RMarkdown or other
situations where you don't control child rendering) or \code{"none"} for no card
styling.}

\item{selector_prefix}{CSS prefix used to target grid elements. This will
change if you're integrating grid with a system that you don't want to use
ids (the \code{"#"} prefix) with because they are not available or are used for
other reasons.}
}
\description{
Adds required hooks to RMarkdown to process \code{gridlayout} chunks and style
document accordingly. Layout will be generated from a chunk identified with
the syntax\preformatted{```\{gridlayout\}
<insert layout table here>
```
}
}
\details{
See \code{vignette("using_with_rmd", package = "gridlayout")} for a more in-depth overview.
}
